requirements.txt
livekit-agents==0.12.2
livekit-plugins-openai==0.10.10
livekit-plugins-deepgram==0.6.14
livekit-plugins-silero==0.7.4
python-dotenv~=1.0

